+++ internal/cli/main.go
@@ -3,6 +3,7 @@ package cli
 import (
 	"fmt"
 	"github.com/rs/zerolog/log"
+	"github.com/thomiceli/opengist/internal/auth/webauthn"
 	"github.com/thomiceli/opengist/internal/config"
 	"github.com/thomiceli/opengist/internal/db"
 	"github.com/thomiceli/opengist/internal/git"
@@ -12,15 +13,17 @@ import (
 	"github.com/thomiceli/opengist/internal/web"
 	"github.com/urfave/cli/v2"
 	"os"
+	"os/signal"
 	"path"
 	"path/filepath"
+	"syscall"
 )
 
 var CmdVersion = cli.Command{
 	Name:  "version",
 	Usage: "Print the version of Opengist",
 	Action: func(c *cli.Context) error {
-		fmt.Println("Opengist v" + config.OpengistVersion)
+		fmt.Println("Opengist " + config.OpengistVersion)
 		return nil
 	},
 }
@@ -29,10 +32,17 @@ var CmdStart = cli.Command{
 	Name:  "start",
 	Usage: "Start Opengist server",
 	Action: func(ctx *cli.Context) error {
+		stopCtx, stop := signal.NotifyContext(ctx.Context, syscall.SIGINT, syscall.SIGTERM)
+		defer stop()
+
 		Initialize(ctx)
-		go web.NewServer(os.Getenv("OG_DEV") == "1").Start()
+
+		go web.NewServer(os.Getenv("OG_DEV") == "1", path.Join(config.GetHomeDir(), "sessions")).Start()
 		go ssh.Start()
-		select {}
+
+		<-stopCtx.Done()
+		shutdown()
+		return nil
 	},
 }
 
@@ -57,7 +67,7 @@ func App() error {
 }
 
 func Initialize(ctx *cli.Context) {
-	fmt.Println("Opengist v" + config.OpengistVersion)
+	fmt.Println("Opengist " + config.OpengistVersion)
 
 	if err := config.InitConfig(ctx.String("config"), os.Stdout); err != nil {
 		panic(err)
@@ -99,8 +109,9 @@ func Initialize(ctx *cli.Context) {
 	if err := os.MkdirAll(filepath.Join(homePath, "custom"), 0755); err != nil {
 		log.Fatal().Err(err).Send()
 	}
-	log.Info().Msg("Database file: " + filepath.Join(homePath, config.C.DBFilename))
-	if err := db.Setup(filepath.Join(homePath, config.C.DBFilename), false); err != nil {
+
+	db.DeprecationDBFilename()
+	if err := db.Setup(config.C.DBUri, false); err != nil {
 		log.Fatal().Err(err).Msg("Failed to initialize database")
 	}
 
@@ -108,12 +119,28 @@ func Initialize(ctx *cli.Context) {
 		log.Fatal().Err(err).Msg("Failed to initialize in memory database")
 	}
 
+	if err := webauthn.Init(config.C.ExternalUrl); err != nil {
+		log.Error().Err(err).Msg("Failed to initialize WebAuthn")
+	}
+
 	if config.C.IndexEnabled {
 		log.Info().Msg("Index directory: " + filepath.Join(homePath, config.C.IndexDirname))
-		if err := index.Open(filepath.Join(homePath, config.C.IndexDirname)); err != nil {
-			log.Fatal().Err(err).Msg("Failed to open index")
-		}
+		index.Init(filepath.Join(homePath, config.C.IndexDirname))
 	}
+}
+
+func shutdown() {
+	log.Info().Msg("Shutting down database...")
+	if err := db.Close(); err != nil {
+		log.Error().Err(err).Msg("Failed to close database")
+	}
+
+	if config.C.IndexEnabled {
+		log.Info().Msg("Shutting down index...")
+		index.Close()
+	}
+
+	log.Info().Msg("Shutdown complete")
 }
 
 func createSymlink(homePath string, configPath string) error {