+++ internal/db/admin_setting.go
@@ -5,15 +5,16 @@ import (
 )
 
 type AdminSetting struct {
-	Key   string `gorm:"uniqueIndex"`
+	Key   string `gorm:"index:,unique"`
 	Value string
 }
 
 const (
-	SettingDisableSignup    = "disable-signup"
-	SettingRequireLogin     = "require-login"
-	SettingDisableLoginForm = "disable-login-form"
-	SettingDisableGravatar  = "disable-gravatar"
+	SettingDisableSignup          = "disable-signup"
+	SettingRequireLogin           = "require-login"
+	SettingAllowGistsWithoutLogin = "allow-gists-without-login"
+	SettingDisableLoginForm       = "disable-login-form"
+	SettingDisableGravatar        = "disable-gravatar"
 )
 
 func GetSetting(key string) (string, error) {
@@ -48,7 +49,7 @@ func UpdateSetting(key string, value string) error {
 }
 
 func setSetting(key string, value string) error {
-	return db.Create(&AdminSetting{Key: key, Value: value}).Error
+	return db.FirstOrCreate(&AdminSetting{Key: key, Value: value}, &AdminSetting{Key: key}).Error
 }
 
 func initAdminSettings(settings map[string]string) error {
@@ -61,4 +62,22 @@ func initAdminSettings(settings map[string]string) error {
 	}
 
 	return nil
+}
+
+type AuthInfo struct{}
+
+func (auth AuthInfo) RequireLogin() (bool, error) {
+	s, err := GetSetting(SettingRequireLogin)
+	if err != nil {
+		return true, err
+	}
+	return s == "1", nil
+}
+
+func (auth AuthInfo) AllowGistsWithoutLogin() (bool, error) {
+	s, err := GetSetting(SettingAllowGistsWithoutLogin)
+	if err != nil {
+		return false, err
+	}
+	return s == "1", nil
 }