+++ internal/web/auth.go
@@ -1,9 +1,10 @@
 package web
 
 import (
+	"bytes"
 	"context"
 	"crypto/md5"
-	"encoding/json"
+	gojson "encoding/json"
 	"errors"
 	"fmt"
 	"github.com/labstack/echo/v4"
@@ -14,8 +15,10 @@ import (
 	"github.com/markbates/goth/providers/gitlab"
 	"github.com/markbates/goth/providers/openidConnect"
 	"github.com/rs/zerolog/log"
+	"github.com/thomiceli/opengist/internal/auth/webauthn"
 	"github.com/thomiceli/opengist/internal/config"
 	"github.com/thomiceli/opengist/internal/db"
+	"github.com/thomiceli/opengist/internal/i18n"
 	"github.com/thomiceli/opengist/internal/utils"
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
@@ -33,8 +36,6 @@ const (
 	OpenIDConnect  = "openid-connect"
 )
 
-var title = cases.Title(language.English)
-
 func register(ctx echo.Context) error {
 	disableSignup := getData(ctx, "DisableSignup")
 	disableForm := getData(ctx, "DisableLoginForm")
@@ -48,8 +49,8 @@ func register(ctx echo.Context) error {
 		}
 	}
 
-	setData(ctx, "title", tr(ctx, "auth.new-account"))
-	setData(ctx, "htmlTitle", "New account")
+	setData(ctx, "title", trH(ctx, "auth.new-account"))
+	setData(ctx, "htmlTitle", trH(ctx, "auth.new-account"))
 	setData(ctx, "disableForm", disableForm)
 	setData(ctx, "disableSignup", disableSignup)
 	setData(ctx, "isLoginPage", false)
@@ -68,30 +69,30 @@ func processRegister(ctx echo.Context) error {
 	}
 
 	if disableSignup == true {
-		return errorRes(403, "Signing up is disabled", nil)
+		return errorRes(403, tr(ctx, "error.signup-disabled"), nil)
 	}
 
 	if getData(ctx, "DisableLoginForm") == true {
-		return errorRes(403, "Signing up via registration form is disabled", nil)
+		return errorRes(403, tr(ctx, "error.signup-disabled-form"), nil)
 	}
 
-	setData(ctx, "title", "New account")
-	setData(ctx, "htmlTitle", "New account")
+	setData(ctx, "title", trH(ctx, "auth.new-account"))
+	setData(ctx, "htmlTitle", trH(ctx, "auth.new-account"))
 
 	sess := getSession(ctx)
 
 	dto := new(db.UserDTO)
 	if err := ctx.Bind(dto); err != nil {
-		return errorRes(400, "Cannot bind data", err)
+		return errorRes(400, tr(ctx, "error.cannot-bind-data"), err)
 	}
 
 	if err := ctx.Validate(dto); err != nil {
-		addFlash(ctx, utils.ValidationMessages(&err), "error")
+		addFlash(ctx, utils.ValidationMessages(&err, getData(ctx, "locale").(*i18n.Locale)), "error")
 		return html(ctx, "auth_form.html")
 	}
 
 	if exists, err := db.UserExists(dto.Username); err != nil || exists {
-		addFlash(ctx, "Username already exists", "error")
+		addFlash(ctx, tr(ctx, "flash.auth.username-exists"), "error")
 		return html(ctx, "auth_form.html")
 	}
 
@@ -126,8 +127,8 @@ func processRegister(ctx echo.Context) error {
 }
 
 func login(ctx echo.Context) error {
-	setData(ctx, "title", tr(ctx, "auth.login"))
-	setData(ctx, "htmlTitle", "Login")
+	setData(ctx, "title", trH(ctx, "auth.login"))
+	setData(ctx, "htmlTitle", trH(ctx, "auth.login"))
 	setData(ctx, "disableForm", getData(ctx, "DisableLoginForm"))
 	setData(ctx, "isLoginPage", true)
 	return html(ctx, "auth_form.html")
@@ -135,7 +136,7 @@ func login(ctx echo.Context) error {
 
 func processLogin(ctx echo.Context) error {
 	if getData(ctx, "DisableLoginForm") == true {
-		return errorRes(403, "Logging in via login form is disabled", nil)
+		return errorRes(403, tr(ctx, "error.login-disabled-form"), nil)
 	}
 
 	var err error
@@ -143,7 +144,7 @@ func processLogin(ctx echo.Context) error {
 
 	dto := &db.UserDTO{}
 	if err = ctx.Bind(dto); err != nil {
-		return errorRes(400, "Cannot bind data", err)
+		return errorRes(400, tr(ctx, "error.cannot-bind-data"), err)
 	}
 	password := dto.Password
 
@@ -154,7 +155,7 @@ func processLogin(ctx echo.Context) error {
 			return errorRes(500, "Cannot get user", err)
 		}
 		log.Warn().Msg("Invalid HTTP authentication attempt from " + ctx.RealIP())
-		addFlash(ctx, "Invalid credentials", "error")
+		addFlash(ctx, tr(ctx, "flash.auth.invalid-credentials"), "error")
 		return redirect(ctx, "/login")
 	}
 
@@ -163,8 +164,19 @@ func processLogin(ctx echo.Context) error {
 			return errorRes(500, "Cannot check for password", err)
 		}
 		log.Warn().Msg("Invalid HTTP authentication attempt from " + ctx.RealIP())
-		addFlash(ctx, "Invalid credentials", "error")
+		addFlash(ctx, tr(ctx, "flash.auth.invalid-credentials"), "error")
 		return redirect(ctx, "/login")
+	}
+
+	// handle MFA
+	var hasMFA bool
+	if hasMFA, err = user.HasMFA(); err != nil {
+		return errorRes(500, "Cannot check for user MFA", err)
+	}
+	if hasMFA {
+		sess.Values["mfaID"] = user.ID
+		saveSession(sess, ctx)
+		return redirect(ctx, "/mfa")
 	}
 
 	sess.Values["user"] = user.ID
@@ -175,10 +187,14 @@ func processLogin(ctx echo.Context) error {
 	return redirect(ctx, "/")
 }
 
+func mfa(ctx echo.Context) error {
+	return html(ctx, "mfa.html")
+}
+
 func oauthCallback(ctx echo.Context) error {
 	user, err := gothic.CompleteUserAuth(ctx.Response(), ctx.Request())
 	if err != nil {
-		return errorRes(400, "Cannot complete user auth: "+err.Error(), err)
+		return errorRes(400, tr(ctx, "error.complete-oauth-login", err.Error()), err)
 	}
 
 	currUser := getUserLogged(ctx)
@@ -187,10 +203,10 @@ func oauthCallback(ctx echo.Context) error {
 		updateUserProviderInfo(currUser, user.Provider, user)
 
 		if err = currUser.Update(); err != nil {
-			return errorRes(500, "Cannot update user "+title.String(user.Provider)+" id", err)
+			return errorRes(500, "Cannot update user "+cases.Title(language.English).String(user.Provider)+" id", err)
 		}
 
-		addFlash(ctx, "Account linked to "+title.String(user.Provider), "success")
+		addFlash(ctx, tr(ctx, "flash.auth.account-linked-oauth", cases.Title(language.English).String(user.Provider)), "success")
 		return redirect(ctx, "/settings")
 	}
 
@@ -198,7 +214,7 @@ func oauthCallback(ctx echo.Context) error {
 	userDB, err := db.GetUserByProvider(user.UserID, user.Provider)
 	if err != nil {
 		if getData(ctx, "DisableSignup") == true {
-			return errorRes(403, "Signing up is disabled", nil)
+			return errorRes(403, tr(ctx, "error.signup-disabled"), nil)
 		}
 
 		if !errors.Is(err, gorm.ErrRecordNotFound) {
@@ -216,7 +232,7 @@ func oauthCallback(ctx echo.Context) error {
 
 		if err = userDB.Create(); err != nil {
 			if db.IsUniqueConstraintViolation(err) {
-				addFlash(ctx, "Username "+user.NickName+" already exists in Opengist", "error")
+				addFlash(ctx, tr(ctx, "flash.auth.username-exists"), "error")
 				return redirect(ctx, "/login")
 			}
 
@@ -246,7 +262,7 @@ func oauthCallback(ctx echo.Context) error {
 
 			body, err := io.ReadAll(resp.Body)
 			if err != nil {
-				addFlash(ctx, "Could not get user keys", "error")
+				addFlash(ctx, tr(ctx, "flash.auth.user-sshkeys-not-retrievable"), "error")
 				log.Error().Err(err).Msg("Could not get user keys")
 			}
 
@@ -262,7 +278,7 @@ func oauthCallback(ctx echo.Context) error {
 				}
 
 				if err = sshKey.Create(); err != nil {
-					addFlash(ctx, "Could not create ssh key", "error")
+					addFlash(ctx, tr(ctx, "flash.auth.user-sshkeys-not-created"), "error")
 					log.Error().Err(err).Msg("Could not create ssh key")
 				}
 			}
@@ -343,36 +359,179 @@ func oauth(ctx echo.Context) error {
 		goth.UseProviders(oidcProvider)
 	}
 
-	currUser := getUserLogged(ctx)
-	if currUser != nil {
-		// Map each provider to a function that checks the relevant ID in currUser
-		providerIDCheckMap := map[string]func() bool{
-			GitHubProvider: func() bool { return currUser.GithubID != "" },
-			GitLabProvider: func() bool { return currUser.GitlabID != "" },
-			GiteaProvider:  func() bool { return currUser.GiteaID != "" },
-			OpenIDConnect:  func() bool { return currUser.OIDCID != "" },
-		}
-
-		// Check if the provider is valid and if the user has a linked ID
-		// Means that the user wants to unlink the account
-		if checkFunc, exists := providerIDCheckMap[provider]; exists && checkFunc() {
-			if err := currUser.DeleteProviderID(provider); err != nil {
-				return errorRes(500, "Cannot unlink account from "+title.String(provider), err)
-			}
-
-			addFlash(ctx, "Account unlinked from "+title.String(provider), "success")
-			return redirect(ctx, "/settings")
-		}
-	}
-
 	ctxValue := context.WithValue(ctx.Request().Context(), gothic.ProviderParamKey, provider)
 	ctx.SetRequest(ctx.Request().WithContext(ctxValue))
 	if provider != GitHubProvider && provider != GitLabProvider && provider != GiteaProvider && provider != OpenIDConnect {
-		return errorRes(400, "Unsupported provider", nil)
+		return errorRes(400, tr(ctx, "error.oauth-unsupported"), nil)
 	}
 
 	gothic.BeginAuthHandler(ctx.Response(), ctx.Request())
 	return nil
+}
+
+func oauthUnlink(ctx echo.Context) error {
+	provider := ctx.Param("provider")
+
+	currUser := getUserLogged(ctx)
+	// Map each provider to a function that checks the relevant ID in currUser
+	providerIDCheckMap := map[string]func() bool{
+		GitHubProvider: func() bool { return currUser.GithubID != "" },
+		GitLabProvider: func() bool { return currUser.GitlabID != "" },
+		GiteaProvider:  func() bool { return currUser.GiteaID != "" },
+		OpenIDConnect:  func() bool { return currUser.OIDCID != "" },
+	}
+
+	if checkFunc, exists := providerIDCheckMap[provider]; exists && checkFunc() {
+		if err := currUser.DeleteProviderID(provider); err != nil {
+			return errorRes(500, "Cannot unlink account from "+cases.Title(language.English).String(provider), err)
+		}
+
+		addFlash(ctx, tr(ctx, "flash.auth.account-unlinked-oauth", cases.Title(language.English).String(provider)), "success")
+		return redirect(ctx, "/settings")
+	}
+
+	return redirect(ctx, "/settings")
+}
+
+func beginWebAuthnBinding(ctx echo.Context) error {
+	credsCreation, jsonWaSession, err := webauthn.BeginBinding(getUserLogged(ctx))
+	if err != nil {
+		return errorRes(500, "Cannot begin WebAuthn registration", err)
+	}
+
+	sess := getSession(ctx)
+	sess.Values["webauthn_registration_session"] = jsonWaSession
+	sess.Options.MaxAge = 5 * 60 // 5 minutes
+	saveSession(sess, ctx)
+
+	return ctx.JSON(200, credsCreation)
+}
+
+func finishWebAuthnBinding(ctx echo.Context) error {
+	sess := getSession(ctx)
+	jsonWaSession, ok := sess.Values["webauthn_registration_session"].([]byte)
+	if !ok {
+		return jsonErrorRes(401, "Cannot get WebAuthn registration session", nil)
+	}
+
+	user := getUserLogged(ctx)
+
+	// extract passkey name from request
+	body, err := io.ReadAll(ctx.Request().Body)
+	if err != nil {
+		return jsonErrorRes(400, "Failed to read request body", err)
+	}
+	ctx.Request().Body.Close()
+	ctx.Request().Body = io.NopCloser(bytes.NewBuffer(body))
+
+	dto := new(db.CrendentialDTO)
+	_ = gojson.Unmarshal(body, &dto)
+
+	if err = ctx.Validate(dto); err != nil {
+		return jsonErrorRes(400, "Invalid request", err)
+	}
+	passkeyName := dto.PasskeyName
+	if passkeyName == "" {
+		passkeyName = "WebAuthn"
+	}
+
+	waCredential, err := webauthn.FinishBinding(user, jsonWaSession, ctx.Request())
+	if err != nil {
+		return jsonErrorRes(403, "Failed binding attempt for passkey", err)
+	}
+
+	if _, err = db.CreateFromCrendential(user.ID, passkeyName, waCredential); err != nil {
+		return jsonErrorRes(500, "Cannot create WebAuthn credential on database", err)
+	}
+
+	delete(sess.Values, "webauthn_registration_session")
+	saveSession(sess, ctx)
+
+	addFlash(ctx, tr(ctx, "flash.auth.passkey-registred", passkeyName), "success")
+	return json(ctx, 200, []string{"OK"})
+}
+
+func beginWebAuthnLogin(ctx echo.Context) error {
+	credsCreation, jsonWaSession, err := webauthn.BeginDiscoverableLogin()
+	if err != nil {
+		return jsonErrorRes(401, "Cannot begin WebAuthn login", err)
+	}
+
+	sess := getSession(ctx)
+	sess.Values["webauthn_login_session"] = jsonWaSession
+	sess.Options.MaxAge = 5 * 60 // 5 minutes
+	saveSession(sess, ctx)
+
+	return json(ctx, 200, credsCreation)
+}
+
+func finishWebAuthnLogin(ctx echo.Context) error {
+	sess := getSession(ctx)
+	sessionData, ok := sess.Values["webauthn_login_session"].([]byte)
+	if !ok {
+		return jsonErrorRes(401, "Cannot get WebAuthn login session", nil)
+	}
+
+	userID, err := webauthn.FinishDiscoverableLogin(sessionData, ctx.Request())
+	if err != nil {
+		return jsonErrorRes(403, "Failed authentication attempt for passkey", err)
+	}
+
+	sess.Values["user"] = userID
+	sess.Options.MaxAge = 60 * 60 * 24 * 365 // 1 year
+
+	delete(sess.Values, "webauthn_login_session")
+	saveSession(sess, ctx)
+
+	return json(ctx, 200, []string{"OK"})
+}
+
+func beginWebAuthnAssertion(ctx echo.Context) error {
+	sess := getSession(ctx)
+
+	ogUser, err := db.GetUserById(sess.Values["mfaID"].(uint))
+	if err != nil {
+		return jsonErrorRes(500, "Cannot get user", err)
+	}
+
+	credsCreation, jsonWaSession, err := webauthn.BeginLogin(ogUser)
+	if err != nil {
+		return jsonErrorRes(401, "Cannot begin WebAuthn login", err)
+	}
+
+	sess.Values["webauthn_assertion_session"] = jsonWaSession
+	sess.Options.MaxAge = 5 * 60 // 5 minutes
+	saveSession(sess, ctx)
+
+	return json(ctx, 200, credsCreation)
+}
+
+func finishWebAuthnAssertion(ctx echo.Context) error {
+	sess := getSession(ctx)
+	sessionData, ok := sess.Values["webauthn_assertion_session"].([]byte)
+	if !ok {
+		return jsonErrorRes(401, "Cannot get WebAuthn assertion session", nil)
+	}
+
+	userId := sess.Values["mfaID"].(uint)
+
+	ogUser, err := db.GetUserById(userId)
+	if err != nil {
+		return jsonErrorRes(500, "Cannot get user", err)
+	}
+
+	if err = webauthn.FinishLogin(ogUser, sessionData, ctx.Request()); err != nil {
+		return jsonErrorRes(403, "Failed authentication attempt for passkey", err)
+	}
+
+	sess.Values["user"] = userId
+	sess.Options.MaxAge = 60 * 60 * 24 * 365 // 1 year
+
+	delete(sess.Values, "webauthn_assertion_session")
+	delete(sess.Values, "mfaID")
+	saveSession(sess, ctx)
+
+	return json(ctx, 200, []string{"OK"})
 }
 
 func logout(ctx echo.Context) error {
@@ -426,7 +585,7 @@ func getAvatarUrlFromProvider(provider string, identifier string) string {
 		}
 
 		var result map[string]interface{}
-		err = json.Unmarshal(body, &result)
+		err = gojson.Unmarshal(body, &result)
 		if err != nil {
 			log.Error().Err(err).Msg("Cannot unmarshal Gitea response body")
 			return ""
@@ -440,4 +599,16 @@ func getAvatarUrlFromProvider(provider string, identifier string) string {
 		return field.(string)
 	}
 	return ""
+}
+
+type ContextAuthInfo struct {
+	context echo.Context
+}
+
+func (auth ContextAuthInfo) RequireLogin() (bool, error) {
+	return getData(auth.context, "RequireLogin") == true, nil
+}
+
+func (auth ContextAuthInfo) AllowGistsWithoutLogin() (bool, error) {
+	return getData(auth.context, "AllowGistsWithoutLogin") == true, nil
 }