+++ internal/web/settings.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"github.com/thomiceli/opengist/internal/config"
 	"github.com/thomiceli/opengist/internal/git"
+	"github.com/thomiceli/opengist/internal/i18n"
 	"github.com/thomiceli/opengist/internal/utils"
 	"os"
 	"path/filepath"
@@ -25,10 +26,17 @@ func userSettings(ctx echo.Context) error {
 		return errorRes(500, "Cannot get SSH keys", err)
 	}
 
+	passkeys, err := db.GetAllCredentialsForUser(user.ID)
+	if err != nil {
+		return errorRes(500, "Cannot get WebAuthn credentials", err)
+	}
+
 	setData(ctx, "email", user.Email)
 	setData(ctx, "sshKeys", keys)
+	setData(ctx, "passkeys", passkeys)
 	setData(ctx, "hasPassword", user.Password != "")
-	setData(ctx, "htmlTitle", "Settings")
+	setData(ctx, "disableForm", getData(ctx, "DisableLoginForm"))
+	setData(ctx, "htmlTitle", trH(ctx, "settings"))
 	return html(ctx, "settings.html")
 }
 
@@ -51,7 +59,7 @@ func emailProcess(ctx echo.Context) error {
 		return errorRes(500, "Cannot update email", err)
 	}
 
-	addFlash(ctx, "Email updated", "success")
+	addFlash(ctx, tr(ctx, "flash.user.email-updated"), "success")
 	return redirect(ctx, "/settings")
 }
 
@@ -70,11 +78,11 @@ func sshKeysProcess(ctx echo.Context) error {
 
 	dto := new(db.SSHKeyDTO)
 	if err := ctx.Bind(dto); err != nil {
-		return errorRes(400, "Cannot bind data", err)
+		return errorRes(400, tr(ctx, "error.cannot-bind-data"), err)
 	}
 
 	if err := ctx.Validate(dto); err != nil {
-		addFlash(ctx, utils.ValidationMessages(&err), "error")
+		addFlash(ctx, utils.ValidationMessages(&err, getData(ctx, "locale").(*i18n.Locale)), "error")
 		return redirect(ctx, "/settings")
 	}
 	key := dto.ToSSHKey()
@@ -83,16 +91,24 @@ func sshKeysProcess(ctx echo.Context) error {
 
 	pubKey, _, _, _, err := ssh.ParseAuthorizedKey([]byte(key.Content))
 	if err != nil {
-		addFlash(ctx, "Invalid SSH key", "error")
+		addFlash(ctx, tr(ctx, "flash.user.invalid-ssh-key"), "error")
 		return redirect(ctx, "/settings")
 	}
 	key.Content = strings.TrimSpace(string(ssh.MarshalAuthorizedKey(pubKey)))
+
+	if exists, err := db.SSHKeyDoesExists(key.Content); exists {
+		if err != nil {
+			return errorRes(500, "Cannot check if SSH key exists", err)
+		}
+		addFlash(ctx, tr(ctx, "settings.ssh-key-exists"), "error")
+		return redirect(ctx, "/settings")
+	}
 
 	if err := key.Create(); err != nil {
 		return errorRes(500, "Cannot add SSH key", err)
 	}
 
-	addFlash(ctx, "SSH key added", "success")
+	addFlash(ctx, tr(ctx, "flash.user.ssh-key-added"), "success")
 	return redirect(ctx, "/settings")
 }
 
@@ -113,7 +129,27 @@ func sshKeysDelete(ctx echo.Context) error {
 		return errorRes(500, "Cannot delete SSH key", err)
 	}
 
-	addFlash(ctx, "SSH key deleted", "success")
+	addFlash(ctx, tr(ctx, "flash.user.ssh-key-deleted"), "success")
+	return redirect(ctx, "/settings")
+}
+
+func passkeyDelete(ctx echo.Context) error {
+	user := getUserLogged(ctx)
+	keyId, err := strconv.Atoi(ctx.Param("id"))
+	if err != nil {
+		return redirect(ctx, "/settings")
+	}
+
+	passkey, err := db.GetCredentialByIDDB(uint(keyId))
+	if err != nil || passkey.UserID != user.ID {
+		return redirect(ctx, "/settings")
+	}
+
+	if err := passkey.Delete(); err != nil {
+		return errorRes(500, "Cannot delete passkey", err)
+	}
+
+	addFlash(ctx, tr(ctx, "flash.auth.passkey-deleted"), "success")
 	return redirect(ctx, "/settings")
 }
 
@@ -122,12 +158,12 @@ func passwordProcess(ctx echo.Context) error {
 
 	dto := new(db.UserDTO)
 	if err := ctx.Bind(dto); err != nil {
-		return errorRes(400, "Cannot bind data", err)
+		return errorRes(400, tr(ctx, "error.cannot-bind-data"), err)
 	}
 	dto.Username = user.Username
 
 	if err := ctx.Validate(dto); err != nil {
-		addFlash(ctx, utils.ValidationMessages(&err), "error")
+		addFlash(ctx, utils.ValidationMessages(&err, getData(ctx, "locale").(*i18n.Locale)), "error")
 		return html(ctx, "settings.html")
 	}
 
@@ -141,7 +177,7 @@ func passwordProcess(ctx echo.Context) error {
 		return errorRes(500, "Cannot update password", err)
 	}
 
-	addFlash(ctx, "Password updated", "success")
+	addFlash(ctx, tr(ctx, "flash.user.password-updated"), "success")
 	return redirect(ctx, "/settings")
 }
 
@@ -150,17 +186,17 @@ func usernameProcess(ctx echo.Context) error {
 
 	dto := new(db.UserDTO)
 	if err := ctx.Bind(dto); err != nil {
-		return errorRes(400, "Cannot bind data", err)
+		return errorRes(400, tr(ctx, "error.cannot-bind-data"), err)
 	}
 	dto.Password = user.Password
 
 	if err := ctx.Validate(dto); err != nil {
-		addFlash(ctx, utils.ValidationMessages(&err), "error")
+		addFlash(ctx, utils.ValidationMessages(&err, getData(ctx, "locale").(*i18n.Locale)), "error")
 		return redirect(ctx, "/settings")
 	}
 
 	if exists, err := db.UserExists(dto.Username); err != nil || exists {
-		addFlash(ctx, "Username already exists", "error")
+		addFlash(ctx, tr(ctx, "flash.auth.username-exists"), "error")
 		return redirect(ctx, "/settings")
 	}
 
@@ -180,6 +216,6 @@ func usernameProcess(ctx echo.Context) error {
 		return errorRes(500, "Cannot update username", err)
 	}
 
-	addFlash(ctx, "Username updated", "success")
+	addFlash(ctx, tr(ctx, "flash.user.username-updated"), "success")
 	return redirect(ctx, "/settings")
 }