+++ internal/web/test/server.go
@@ -3,13 +3,6 @@ package test
 import (
 	"errors"
 	"fmt"
-	"github.com/rs/zerolog/log"
-	"github.com/stretchr/testify/require"
-	"github.com/thomiceli/opengist/internal/config"
-	"github.com/thomiceli/opengist/internal/db"
-	"github.com/thomiceli/opengist/internal/git"
-	"github.com/thomiceli/opengist/internal/memdb"
-	"github.com/thomiceli/opengist/internal/web"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -21,7 +14,17 @@ import (
 	"strconv"
 	"strings"
 	"testing"
+
+	"github.com/rs/zerolog/log"
+	"github.com/stretchr/testify/require"
+	"github.com/thomiceli/opengist/internal/config"
+	"github.com/thomiceli/opengist/internal/db"
+	"github.com/thomiceli/opengist/internal/git"
+	"github.com/thomiceli/opengist/internal/memdb"
+	"github.com/thomiceli/opengist/internal/web"
 )
+
+var databaseType string
 
 type testServer struct {
 	server        *web.Server
@@ -30,7 +33,7 @@ type testServer struct {
 
 func newTestServer() (*testServer, error) {
 	s := &testServer{
-		server: web.NewServer(true),
+		server: web.NewServer(true, path.Join(config.GetHomeDir(), "tmp", "sessions")),
 	}
 
 	go s.start()
@@ -106,7 +109,7 @@ func structToURLValues(s interface{}) url.Values {
 	for i := 0; i < rValue.NumField(); i++ {
 		field := rValue.Type().Field(i)
 		tag := field.Tag.Get("form")
-		if tag != "" {
+		if tag != "" || field.Anonymous {
 			if field.Type.Kind() == reflect.Int {
 				fieldValue := rValue.Field(i).Int()
 				v.Add(tag, strconv.FormatInt(fieldValue, 10))
@@ -114,6 +117,12 @@ func structToURLValues(s interface{}) url.Values {
 				fieldValue := rValue.Field(i).Interface().([]string)
 				for _, va := range fieldValue {
 					v.Add(tag, va)
+				}
+			} else if field.Type.Kind() == reflect.Struct {
+				for key, val := range structToURLValues(rValue.Field(i).Interface()) {
+					for _, vv := range val {
+						v.Add(key, vv)
+					}
 				}
 			} else {
 				fieldValue := rValue.Field(i).String()
@@ -125,6 +134,17 @@ func structToURLValues(s interface{}) url.Values {
 }
 
 func setup(t *testing.T) {
+	var databaseDsn string
+	databaseType = os.Getenv("OPENGIST_TEST_DB")
+	switch databaseType {
+	case "sqlite":
+		databaseDsn = "file::memory:"
+	case "postgres":
+		databaseDsn = "postgres://postgres:opengist@localhost:5432/opengist_test"
+	case "mysql":
+		databaseDsn = "mysql://root:opengist@localhost:3306/opengist_test"
+	}
+
 	_ = os.Setenv("OPENGIST_SKIP_GIT_HOOKS", "1")
 
 	err := config.InitConfig("", io.Discard)
@@ -142,14 +162,18 @@ func setup(t *testing.T) {
 	homePath := config.GetHomeDir()
 	log.Info().Msg("Data directory: " + homePath)
 
-	err = os.MkdirAll(filepath.Join(homePath, "sessions"), 0755)
+	err = os.MkdirAll(filepath.Join(homePath, "tmp", "sessions"), 0755)
 	require.NoError(t, err, "Could not create sessions directory")
 
 	err = os.MkdirAll(filepath.Join(homePath, "tmp", "repos"), 0755)
 	require.NoError(t, err, "Could not create tmp repos directory")
 
-	err = db.Setup("file::memory:", true)
+	err = db.Setup(databaseDsn, true)
 	require.NoError(t, err, "Could not initialize database")
+
+	if err != nil {
+		log.Fatal().Err(err).Msg("Could not initialize database")
+	}
 
 	err = memdb.Setup()
 	require.NoError(t, err, "Could not initialize in memory database")
@@ -161,14 +185,20 @@ func setup(t *testing.T) {
 func teardown(t *testing.T, s *testServer) {
 	s.stop()
 
-	err := db.Close()
-	require.NoError(t, err, "Could not close database")
+	//err := db.Close()
+	//require.NoError(t, err, "Could not close database")
 
-	err = os.RemoveAll(path.Join(config.GetHomeDir(), "tests"))
+	err := os.RemoveAll(path.Join(config.GetHomeDir(), "tests"))
 	require.NoError(t, err, "Could not remove repos directory")
 
 	err = os.RemoveAll(path.Join(config.GetHomeDir(), "tmp", "repos"))
 	require.NoError(t, err, "Could not remove repos directory")
+
+	err = os.RemoveAll(path.Join(config.GetHomeDir(), "tmp", "sessions"))
+	require.NoError(t, err, "Could not remove repos directory")
+
+	err = db.TruncateDatabase()
+	require.NoError(t, err, "Could not truncate database")
 
 	// err = os.RemoveAll(path.Join(config.C.OpengistHome, "testsindex"))
 	// require.NoError(t, err, "Could not remove repos directory")